Fixe Interface + Add deprecation
