+++ monailabel/endpoints/scoring.py
@@ -73,7 +73,7 @@ async def api_status(
     return status(all, check_if_running)
 
 
-@router.post("/", summary="Run All Scoring Tasks")
+@router.post("/", summary="Run All Scoring Tasks", deprecated=True)
 async def api_run(
     params: Optional[dict] = None,
     run_sync: Optional[bool] = False,
