+++ monailabel/endpoints/train.py
@@ -76,7 +76,7 @@ async def api_status(
     return status(all, check_if_running)
 
 
-@router.post("/", summary="Run All Training Tasks")
+@router.post("/", summary="Run All Training Tasks", deprecated=True)
 async def api_run(
     params: Optional[dict] = None,
     run_sync: Optional[bool] = False,
