+++ monailabel/interfaces/config.py
@@ -13,7 +13,7 @@ import logging
 from abc import ABCMeta, abstractmethod
 from typing import Any, Dict, Optional, Union
 
-from monailabel.interfaces.tasks.infer import InferTask
+from monailabel.interfaces.tasks.infer_v2 import InferTask
 from monailabel.interfaces.tasks.scoring import ScoringMethod
 from monailabel.interfaces.tasks.strategy import Strategy
 from monailabel.interfaces.tasks.train import TrainTask
