+++ sample-apps/pathology/lib/infers/nuclick.py
@@ -27,7 +27,8 @@ from monai.transforms import (
 )
 from monai.utils import ensure_tuple
 
-from monailabel.interfaces.tasks.infer import InferTask, InferType
+from monailabel.interfaces.tasks.infer_v2 import InferType
+from monailabel.tasks.infer.basic_infer import BasicInferTask
 from monailabel.transform.post import FindContoursd
 from monailabel.transform.writer import PolygonWriter
 
@@ -63,7 +64,7 @@ class ConvertInteractiveClickSignals(MapTransform):
         return data
 
 
-class NuClick(InferTask):
+class NuClick(BasicInferTask):
     """
     This provides Inference Engine for pre-trained NuClick segmentation (UNet) model.
     """
