+++ sample-apps/radiology/lib/configs/localization_spine.py
@@ -18,7 +18,7 @@ import lib.trainers
 from monai.networks.nets import UNet
 
 from monailabel.interfaces.config import TaskConfig
-from monailabel.interfaces.tasks.infer import InferTask
+from monailabel.interfaces.tasks.infer_v2 import InferTask
 from monailabel.interfaces.tasks.train import TrainTask
 from monailabel.utils.others.generic import download_file, strtobool
