+++ sample-apps/radiology/lib/infers/segmentation.py
@@ -23,11 +23,12 @@ from monai.transforms import (
     Spacingd,
 )
 
-from monailabel.interfaces.tasks.infer import InferTask, InferType
+from monailabel.interfaces.tasks.infer_v2 import InferType
+from monailabel.tasks.infer.basic_infer import BasicInferTask
 from monailabel.transform.post import Restored
 
 
-class Segmentation(InferTask):
+class Segmentation(BasicInferTask):
     """
     This provides Inference Engine for pre-trained segmentation (UNet) model over MSD Dataset.
     """
