+++ sample-apps/radiology/lib/infers/segmentation_spleen.py
@@ -22,11 +22,12 @@ from monai.transforms import (
     Spacingd,
 )
 
-from monailabel.interfaces.tasks.infer import InferTask, InferType
+from monailabel.interfaces.tasks.infer_v2 import InferType
+from monailabel.tasks.infer.basic_infer import BasicInferTask
 from monailabel.transform.post import Restored
 
 
-class SegmentationSpleen(InferTask):
+class SegmentationSpleen(BasicInferTask):
     """
     This provides Inference Engine for pre-trained spleen segmentation (UNet) model over MSD Dataset.
     """
