+++ sample-apps/radiology/main.py
@@ -23,7 +23,7 @@ import monailabel
 from monailabel.interfaces.app import MONAILabelApp
 from monailabel.interfaces.config import TaskConfig
 from monailabel.interfaces.datastore import Datastore
-from monailabel.interfaces.tasks.infer import InferTask
+from monailabel.interfaces.tasks.infer_v2 import InferTask
 from monailabel.interfaces.tasks.scoring import ScoringMethod
 from monailabel.interfaces.tasks.strategy import Strategy
 from monailabel.interfaces.tasks.train import TrainTask
@@ -128,7 +128,6 @@ class MyApp(MONAILabelApp):
                     lamda=1.0,
                     sigma=0.1,
                     num_bins=64,
-                    labels=task_config.labels,
                 ),
                 "GMM+GraphCut": GMMBasedGraphCut(
                     intensity_range=(-300, 200, 0.0, 1.0, True),
@@ -136,7 +135,6 @@ class MyApp(MONAILabelApp):
                     lamda=5.0,
                     sigma=0.5,
                     num_mixtures=20,
-                    labels=task_config.labels,
                 ),
             }
         )
